Migrate models to separate module
