+++ pom.xml
@@ -10,6 +10,7 @@
   <modules>
     <module>server</module>
     <module>client</module>
+    <module>wilson-models</module>
   </modules>
 
   <properties>
