+++ server/src/main/java/com/github/nhirakawa/wilson/server/WilsonRunner.java
@@ -13,12 +13,12 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.fasterxml.jackson.core.type.TypeReference;
+import com.github.nhirakawa.wilson.models.ClusterMember;
 import com.github.nhirakawa.wilson.server.config.ConfigPath;
 import com.github.nhirakawa.wilson.server.config.ConfigValidator;
 import com.github.nhirakawa.wilson.server.dagger.DaggerWilsonServerComponent;
 import com.github.nhirakawa.wilson.server.dagger.WilsonDaggerModule;
 import com.github.nhirakawa.wilson.server.jackson.ObjectMapperWrapper;
-import com.github.nhirakawa.wilson.server.models.ClusterMember;
 import com.google.common.base.Throwables;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Sets;
@@ -30,92 +30,93 @@ import com.typesafe.config.ConfigRenderOptions;
 
 public class WilsonRunner {
 
-  private static final Logger LOG = LoggerFactory.getLogger(WilsonRunner.class);
+	private static final Logger LOG = LoggerFactory.getLogger(WilsonRunner.class);
 
-  public static void main(String... args) throws IOException {
-    printBanner();
+	public static void main(String... args) throws IOException {
+		printBanner();
 
-    Config config = ConfigFactory.load();
-    ConfigValidator.validate(config);
-    boolean isLocalCluster = config.getBoolean(ConfigPath.WILSON_LOCAL_CLUSTER.getPath());
+		Config config = ConfigFactory.load();
+		ConfigValidator.validate(config);
+		boolean isLocalCluster = config.getBoolean(ConfigPath.WILSON_LOCAL_CLUSTER.getPath());
 
-    if (isLocalCluster) {
-      runLocalCluster(config);
-    } else {
-      runLocalMember(config);
-    }
-  }
+		if (isLocalCluster) {
+			runLocalCluster(config);
+		} else {
+			runLocalMember(config);
+		}
+	}
 
-  private static void runLocalCluster(Config config) throws IOException {
-    Set<ClusterMember> clusterMembers = ObjectMapperWrapper.instance().readValue(
-        config.getList(ConfigPath.WILSON_CLUSTER_ADDRESSES.getPath()).render(ConfigRenderOptions.concise()),
-        new TypeReference<Set<ClusterMember>>() {}
-    );
+	private static void runLocalCluster(Config config) throws IOException {
+		Set<ClusterMember> clusterMembers = ObjectMapperWrapper.instance().readValue(
+				config.getList(ConfigPath.WILSON_CLUSTER_ADDRESSES.getPath()).render(ConfigRenderOptions.concise()),
+				new TypeReference<Set<ClusterMember>>() {}
+		);
 
-    ExecutorService executorService = getExecutorService();
+		ExecutorService executorService = getExecutorService();
 
-    for (ClusterMember clusterMember : clusterMembers) {
-      executorService.execute(() -> runMember(clusterMember, clusterMembers, config));
-    }
-  }
+		for (ClusterMember clusterMember : clusterMembers) {
+			executorService.execute(() -> runMember(clusterMember, clusterMembers, config));
+		}
+	}
 
-  private static void runLocalMember(Config config) throws IOException {
-    ClusterMember localClusterMember = ObjectMapperWrapper.instance().readValue(
-        config.getObject(ConfigPath.WILSON_LOCAL_ADDRESS.getPath()).render(ConfigRenderOptions.concise()),
-        new TypeReference<ClusterMember>() {}
-    );
+	private static void runLocalMember(Config config) throws IOException {
+		ClusterMember localClusterMember = ObjectMapperWrapper.instance().readValue(
+				config.getObject(ConfigPath.WILSON_LOCAL_ADDRESS.getPath()).render(ConfigRenderOptions.concise()),
+				new TypeReference<ClusterMember>() {}
+		);
 
-    Set<ClusterMember> clusterMembers = ObjectMapperWrapper.instance().readValue(
-        config.getList(ConfigPath.WILSON_CLUSTER_ADDRESSES.getPath()).render(ConfigRenderOptions.concise()),
-        new TypeReference<Set<ClusterMember>>() {}
-    );
+		Set<ClusterMember> clusterMembers = ObjectMapperWrapper.instance().readValue(
+				config.getList(ConfigPath.WILSON_CLUSTER_ADDRESSES.getPath()).render(ConfigRenderOptions.concise()),
+				new TypeReference<Set<ClusterMember>>() {}
+		);
 
-    runMember(localClusterMember, clusterMembers, config);
-  }
+		runMember(localClusterMember, clusterMembers, config);
+	}
 
-  private static void runMember(ClusterMember clusterMember, Set<ClusterMember> clusterMembers, Config config) {
-    Config configWithLocalMember = ConfigFactory.parseMap(
-        ImmutableMap.of(
-            "wilson.local.host", clusterMember.getHost(),
-            "wilson.local.port", clusterMember.getPort()
-        )
-    )
-        .withFallback(config);
+	private static void runMember(ClusterMember clusterMember, Set<ClusterMember> clusterMembers, Config config) {
+		Config configWithLocalMember = ConfigFactory.parseMap(
+				ImmutableMap.of(
+						"wilson.local.host", clusterMember.getHost(),
+						"wilson.local.port", clusterMember.getPort()
+				)
+		)
+				.withFallback(config);
 
-    Set<ClusterMember> clusterWithoutLocalMember = Sets.difference(
-        clusterMembers,
-        Collections.singleton(clusterMember)
-    );
+		Set<ClusterMember> clusterWithoutLocalMember = Sets.difference(
+				clusterMembers,
+				Collections.singleton(clusterMember)
+		);
 
-    try {
-      DaggerWilsonServerComponent.builder()
-          .wilsonDaggerModule(new WilsonDaggerModule(configWithLocalMember, clusterMember, clusterWithoutLocalMember))
-          .build()
-          .create()
-          .start();
-    } catch (IOException | InterruptedException e) {
-      LOG.error("Could not bootstrap {}", clusterMember, e);
-      Throwables.throwIfUnchecked(e);
-      throw new RuntimeException(e);
-    }
-  }
+		try {
+			DaggerWilsonServerComponent.builder()
+					.wilsonDaggerModule(new WilsonDaggerModule(configWithLocalMember, clusterMember, clusterWithoutLocalMember))
+					.build()
+					.create()
+					.start();
+		} catch (IOException | InterruptedException e) {
+			LOG.error("Could not bootstrap {}", clusterMember, e);
+			Throwables.throwIfUnchecked(e);
+			throw new RuntimeException(e);
+		}
+	}
 
-  private static ExecutorService getExecutorService() {
-    ThreadFactory threadFactory = new ThreadFactoryBuilder()
-        .setNameFormat("wilson-runner-%s")
-        .setDaemon(false)
-        .setUncaughtExceptionHandler((t, e) -> LOG.error("Uncaught exception in thread {}", t, e))
-        .build();
+	private static ExecutorService getExecutorService() {
+		ThreadFactory threadFactory = new ThreadFactoryBuilder()
+				.setNameFormat("wilson-runner-%s")
+				.setDaemon(false)
+				.setUncaughtExceptionHandler((t, e) -> LOG.error("Uncaught exception in thread {}", t, e))
+				.build();
 
-    return Executors.newFixedThreadPool(10, threadFactory);
-  }
+		return Executors.newFixedThreadPool(10, threadFactory);
+	}
 
-  private static void printBanner() {
-    try {
-      URL url = Resources.getResource("banner.txt");
-      String banner = Resources.toString(url, StandardCharsets.UTF_8);
-      LOG.info("\n{}\n", banner);
-    } catch (IllegalArgumentException | IOException ignored) {
-    }
-  }
+	private static void printBanner() {
+		try {
+			URL url = Resources.getResource("banner.txt");
+			String banner = Resources.toString(url, StandardCharsets.UTF_8);
+			LOG.info("\n{}\n", banner);
+		} catch (IllegalArgumentException | IOException ignored) {
+		}
+	}
+
 }
