--- server/src/main/java/com/github/nhirakawa/wilson/server/models/messages/LocalWilsonMessage.java
@@ -1,4 +1,0 @@
-package com.github.nhirakawa.wilson.server.models.messages;
-
-public interface LocalWilsonMessage {
-}
