+++ wilson-models/src/main/java/com/github/nhirakawa/wilson/models/AppendEntriesResult.java
@@ -1,0 +1,9 @@
+package com.github.nhirakawa.wilson.models;
+
+public enum AppendEntriesResult {
+
+	SUCCESS,
+	OUTDATED_TERM,
+	INVALID_LAST_LOG_INDEX,
+
+}
