+++ wilson-models/src/main/java/com/github/nhirakawa/wilson/models/LogItemModel.java
@@ -1,10 +1,10 @@
-package com.github.nhirakawa.wilson.server.models;
+package com.github.nhirakawa.wilson.models;
 
 import java.util.List;
 
 import org.immutables.value.Value.Immutable;
 
-import com.github.nhirakawa.wilson.server.models.style.WilsonStyle;
+import com.github.nhirakawa.wilson.models.style.WilsonStyle;
 
 @Immutable
 @WilsonStyle
