+++ wilson-models/src/main/java/com/github/nhirakawa/wilson/models/messages/AppendEntriesRequestModel.java
@@ -1,11 +1,11 @@
-package com.github.nhirakawa.wilson.server.models.messages;
+package com.github.nhirakawa.wilson.models.messages;
 
 import java.util.List;
 
 import org.immutables.value.Value.Immutable;
 
-import com.github.nhirakawa.wilson.server.models.ClusterMember;
-import com.github.nhirakawa.wilson.server.models.style.WilsonStyle;
+import com.github.nhirakawa.wilson.models.ClusterMember;
+import com.github.nhirakawa.wilson.models.style.WilsonStyle;
 import com.google.protobuf.ByteString;
 
 @WilsonStyle
