+++ wilson-models/src/main/java/com/github/nhirakawa/wilson/models/messages/AppendEntriesResponseModel.java
@@ -1,0 +1,15 @@
+package com.github.nhirakawa.wilson.models.messages;
+
+import org.immutables.value.Value.Immutable;
+
+import com.github.nhirakawa.wilson.models.AppendEntriesResult;
+import com.github.nhirakawa.wilson.models.style.WilsonStyle;
+
+@WilsonStyle
+@Immutable
+public interface AppendEntriesResponseModel {
+
+  long getTerm();
+  AppendEntriesResult getResult();
+
+}
