+++ wilson-models/src/main/java/com/github/nhirakawa/wilson/models/messages/HeartbeatTimeoutMessageModel.java
@@ -1,0 +1,20 @@
+package com.github.nhirakawa.wilson.models.messages;
+
+import java.time.Instant;
+
+import org.immutables.value.Value;
+
+import com.github.nhirakawa.wilson.models.style.WilsonStyle;
+
+@Value.Immutable
+@WilsonStyle
+public interface HeartbeatTimeoutMessageModel {
+
+	@Value.Default
+	default Instant getTimestamp() {
+		return Instant.now();
+	}
+
+	long getHeartbeatTimeout();
+
+}
