+++ wilson-models/src/main/java/com/github/nhirakawa/wilson/models/messages/LocalWilsonMessage.java
@@ -1,0 +1,4 @@
+package com.github.nhirakawa.wilson.models.messages;
+
+public interface LocalWilsonMessage {
+}
