+++ wilson-models/src/main/java/com/github/nhirakawa/wilson/models/messages/UuidWilsonMessageModel.java
@@ -1,11 +1,11 @@
-package com.github.nhirakawa.wilson.server.models.messages;
+package com.github.nhirakawa.wilson.models.messages;
 
 import java.util.UUID;
 
 import org.immutables.value.Value;
 import org.immutables.value.Value.Immutable;
 
-import com.github.nhirakawa.wilson.server.models.style.WilsonStyle;
+import com.github.nhirakawa.wilson.models.style.WilsonStyle;
 
 @Immutable
 @WilsonStyle
