+++ wilson-models/src/main/java/com/github/nhirakawa/wilson/models/messages/VoteRequestModel.java
@@ -1,9 +1,9 @@
-package com.github.nhirakawa.wilson.server.models.messages;
+package com.github.nhirakawa.wilson.models.messages;
 
 import org.immutables.value.Value.Immutable;
 
-import com.github.nhirakawa.wilson.server.models.ClusterMember;
-import com.github.nhirakawa.wilson.server.models.style.WilsonStyle;
+import com.github.nhirakawa.wilson.models.ClusterMember;
+import com.github.nhirakawa.wilson.models.style.WilsonStyle;
 
 @WilsonStyle
 @Immutable
